backend/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.32.0
pydantic>=2.10.0
pydantic-settings>=2.6.0
python-multipart>=0.0.18
sse-starlette>=2.1.0
httpx>=0.28.0
firebase-admin>=6.5.0
stripe>=11.0.0
redis>=5.0.0
aiosqlite>=0.20.0
numpy>=1.26.0
sqlalchemy>=2.0.0
boto3>=1.35.0
pyjwt>=2.9.0
passlib>=1.7.4
python-jose>=3.3.0
cryptography>=42.0.0
pytest>=8.0.0
pytest-asyncio>=0.24.0
pytest-cov>=5.0.0
jinja2>=3.1.0
apscheduler>=3.10.0
livekit-server-sdk>=0.7.0
chromadb>=0.4.22
sentence-transformers>=2.2.0
pytrends>=4.9.0
